Overhaul of WoWAPI unit testing module.
The WoWAPI module tries to provide enough of the WoW environment to be
able to load addon files and do unit testing or integration testing of
addon components.

WoWAPI contains fake implementations of some libraries that are used
by Ovale.

Example usage:

    local state = {
        class = "DRUID",
        level = 90,
    }
    dofile("WoWAPI.lua")
    WoWAPI:Initialize("Ovale", state)
    WoWAPI:ExportSymbols()
    WoWAPI:LoadTOC("Ovale.toc")

This example loads all of the Ovale addon's files in the order listed in
its TOC file.

git-svn-id: svn://svn.curseforge.net/wow/ovale/mainline/trunk@1532 d5049fe3-3747-40f7-a4b5-f36d6801af5f