Change Enemies() to default to only counting enemies you have attacked.
To get the old behavior of counting all enemies that are hostile to the
player's group, then use "Enemies(any=1)".

This more closely matches expectations of script writers that want to do
different actions when faced with differing numbers of enemies -- it's
sensible to count adds only when they are in range of your attacks.

This fixes ticket 323 by @hoogoon.