Properly account for combo points generated from Anticipation stacks.
On a rogue talented into Anticipation, combo points generated when at max
combo points are added to stacks of Anticipation, and when a finisher is
used, any stacks of Anticipation are converted into combo points on the
target.  However, there is a slight delay between the combo points being
consumed and the Anticipation stacks are converted into combot points on
the target.

"Pre-adjust" the combo point total to include the Anticipation stacks when
an offensive finisher is used. Also, adjust the combo points in the
simulator accordingly after a spell is cast.

Add a debug option for the OvaleComboPoint module.