Move racial actions into the class-specific common scripts.
Some racials are class-specific, so define a class-specific
UseRacialActions() function that has a much narrower list of available
racials.

Also move UseRacialInterruptActions() into the bodies of the
class-specific Interrupt() functions and, again, narrow the list of
available racials to the ones from the races allowed to be that class.

Fix the spell ID of Blood Fury on orcs, which has a different ID based on
the class.