Put all script condition definitions into a single file.
This reverses a design decision to split up conditions into smaller files
that ended up making this addon harder to update for users.  Keeping all
conditions in a single file reduces the likelihood of files being added or
removed, which allows for updating the addon without needing a full
restart of the WoW client.

This does make the conditions.lua file harder to maintain, but a clean
internal code layout should alleviate this concern.