Allow specifying a cooldown time for buffs.
This cooldown time is how long to expect between when the buff is gained
and can be used to note the internal cooldown of a trinket that applies a
buff.

Decorate trinket buffs with approximate internal cooldown times.  The
syntax is "buff_cd=N" where N is the number of seconds.

Add a BuffCooldown(id) script condition to return the remaining time
before the buff can be gained again.