Version-control the localization data.
Create a new top-level Localization.lua file that holds the table of
localized strings used within Ovale.

The new "localization.lua" script will fetch the localized strings for
known locales from the Ovale CurseForge site and directly append the data
to the top-level Localization.lua file.

Remove the various separated locale data files since they are now replaced
by Localization.lua.  Having the locale data in a single file loaded by
the WoW client makes it easier to update the addon when new locales are
added.  This change also removes a dependency on AceLocale-3.0.

This also makes it simpler to create a "no-lib" version of this addon
since all of the localized strings are already in a checkout of the
repository.

This also brings the localized strings under version control so that older
version can be restored if necessary.