Allow a new icon parameter "enemies=N" to pretend N enemies are present.
Add state to OvaleEnemies that may be manipulated to set the number of
enemies we pretend are present.  The property "enemies" may be changed
after the state machine is initialized.

When processing each icon node of the script, see if "enemies=N" is given
as a parameter and use that value to set the "enemies" property of the
state machine.

Modify the Enemies() script condition to always return the "enemies"
property of the state machine if it is set, or else fall back to the true
values of active or tagged enemies.

This allows for a single script function to be used for both a
single-target rotation icon and an AoE rotation icon via:

    AddIcon enemies=1 { DoRotation() }
    AddIcon enemies=5 { DoRotation() }

where DoRotation() is a function that can conditionally suggest different
actions based on the value of Enemies().