Fix the way variables are handled using GetState/SetState in scripts.
Internally, SetState() is a special type of action that is "cast"
depending on whether the time span for the action includes "now" or if
it's in the future.

Define a new method "GetAction" for OvaleBestAction that takes a master
node and emits a <timespan, priority, element> triplet for the computed
best action.  GetAction() automatically deals with SetState() actions
during the computation.

This fixes the fire mage script getting stuck when using GetState() and
SetState().