Simpilify processing of changes to the addon options.
Use inherited getter/setter functions instead of using the same template
across all options.

Trigger an event when the options are changed so that the icon and menu
options frames are re-drawn.  This de-couples the toggling of options from
forcing a frame refresh in the same execution thread as the option being
toggled.