Reverse the semantics of Enemies() to restore the old interpretation.
By default, Enemies() will now count the total number of enemies, not just
the ones that are tagged by the player.  To show only the number of tagged
enemies, use Enemies(tagged=1).

This makes Enemies() more useful now that we can restrict an icon to show
the rotation for a specific number of enemies via the enemies=N con
parameter.