Shaman: Use the spell ID from the spell on the action bar for Ascendance.
Ascendance has ID 114049 on the action bar, but the spell cooldown is
tracked via the spec-specific variation of Ascendance.  Modify the shaman
spell information for Ascendance so that the "ascendance_*" names all
refer to the correct spell and force them to use the correct spell to
track the cooldown.

This also fixes showing the keybinding for Ascendance on the Ovale icon.

This fixes ticket 417 by @baksas.