Warrior: Move "offgcd" icon to the far left position.
The default icon layout is now:

    [offgcd] [shortcd] [main] [aoe] [cd]

where the AoE icon can be toggled.  This change makes the shortcd-to-cd
icon block like all other classes and specializations, and simply appends
the offgcd icon to the leftmost position where it can "flash" as a
reminder to use the offgcd ability.

Also, make the offgcd icon small to de-emphasize it relative to the
rotation icons.