Add new method OvalePower:PowerCost() that returns the cost of a spell.
The Blizzard API function GetSpellInfo() in patch 6.0 no longer returns
spell cost and power type information, so add a new method to OvalePower
to return this lost information.  The information is gleaned from scanning
the tooltip for a spell.

The tooltip-scanning uses a method that should correctly return the cost
and power type for any locale.