GetSpellInfo() changed its return values in patch 6.0.
New return values are:

    name, subName, icon, castingTime, minRange, maxRange

Adapt GetSpellInfo() usage to the new return values.  Use the
OvaleSpellBook:GetSpellName() wrapper method in place of the most common
usage to get the name of a spell.  Also add a new method
OvaleSpellBook:GetCastTime() to wrap using GetSpellInfo() to get the
casting time of a spell.