Split out LDB data broker object into a separate module.
Structure the module so that it does nothing if LibDataBroker isn't
present, and use .pkgmeta to pull LDB into the final package instead of
hard-embedding LDB into the addon repository.

This allows us to easily remove LDB support if the licensing terms change
and permission to use LDB is revoked.