Update OvaleSteadyFocus for patch 6.0.
Steady Focus is now a level 60 talent for all hunters.  Steady Focus is
now gained if two Steady Shot (Cobra Shot or Focusing Shot also
acceptable) abilities are used back-to-back.  Track the use of the ability
using UNIT_SPELLCAST_SUCCEEDED instead of CLEU.

Add a debugging option for OvaleSteadyFocus that can be toggled in the
"debug" options panel.