Implement new state method "IsUsableSpell()" in OvaleSpellBook.
The state method IsUsableSpell() returns whether the spell can be cast
given the restrictions set out in the spell information and whether there
are enough secondary resources if necessary.  The implementation is taken
from the GetActionSpellInfo() private method in OvaleBestAction, which has
now been modified to use the state method instead.

Also add some debugging and profiling support for OvaleSpellBook.