Add "GetSpellInfo()" method to OvaleSpellBook.
This method wraps the Blizzard API function and first tries to convert the
given spell ID into a spellbook index, or else falls back to using the
spell ID as the parameter to the Blizzard GetSpellInfo().  This makes sure
that the spell information for any spell replacements in the spellbook are
returned, as is common when a proc happens, or a spell that is different
per specialization.