Allow "learn=1" as a SpellInfo() parameter to force learning the spell.
In patch 6.0, Blizzard removed many spells learned along the way while
leveling that simply modified other spells.  However, these removed spells
were useful as conditions for various spell effects.

This change allows "learn=1" to force adding the spell to the spellbook so
that it may be used as a test condition, regardless of whether the spell
is truly in the in-game spellbook.