Reorganize options to allow for an "enable icons" toggle.
The Visibility options group is now a subgroup of the "Icon Group" options
group, which has options for the bar of icons for the script.  There is a
master option "Enabled" which toggles whether the icon bar is
enabled/visible, and which triggers enabling the other options.

This allows for hiding the icons while still having the script processed
by Ovale.