Re-parent the options panel so it has independent visibility.
The "contents" frame containing the script options widgets had the Ovale
frame as it parent, so it was hidden when that frame was hidden.  Instead,
set the updateFrame as the parent so that hiding/showing the Ovale frame
only hides the icons and the green "move" bar.

This allows for hiding the icons but still allowing to toggle the options
panel via the minimap button.