Change "6.0.2" to "6.0" in scripts.
The versions of WoW for which the scripts apply are the major release 6.0,
all teeny versions.  Drop the last digit of the version number to make it
more apparent, and to forestall questions about when a script will be
updated for the latest teeny version update of WoW.

Leave the teeny versions on the spell data scripts to indicate from which
patch the data was taken.  The spell data can change substantially between
teeny updates, so keep that number as documentation.