Fix Blood DK rotation's Outbreak and Blood Boil usage.
The SimulationCraft action list upon which this script was based is
incorrect.  Fix the conditions for using Outbreak so that it is suggested
when diseases are within 8 seconds of falling off, not when the diseases
have more than 8 seconds left.  Also remove the Blood Boil spam, which was
consuming all blood and death runes and starving the tank of resources to
Death Strike.

This fixes ticket 432.