Use non-recursive post-order traversal when computing the best action.
For a given node, the "postOrder" is an array of nodes with the following
properties:

(1) Odd indices hold nodes in post-order traversal order.
(2) Each even index holds the parent of the node held by the preceding odd
    index.
(3) The last node of the array is the given node.

Add PostOrderCompute() method to use the post-order traversal information.

If the postOrder list is present in a node, then it computes each node in
the list and returns the result of the final computation (the last node in
the postOrder list is the node itself).

Fixes ticket 438 by @jlam.