Update auras caused by the player on non-unitID units.
Catch CLEU events for auras caused by the player on units that don't map
to a unit ID.  These auras are applied at the current aura age, and when
the player targets the unit, the auras are all re-scanned.

This allows Ovale to track auras that are spread by player spells, e.g.,
diseases spread by Blood Boil.