Who knew that frame anchors could mysteriously disappear?
Instead of searching for a CENTER anchor for the frame, directly compute
the X- and Y-offsets of the icon bar frame from its parent frame when the
frame has been moved.

Also fix the offsets to be saved if the values are directly manipulated in
the Ovale configuration panel.

Fixes ticket 445 by @thetravellor.