When looping due to setting a state variable, invalidate the result cache.
When SetState() is encountered in the script, OvaleState:GetAction() sets
the state and re-evaluates the script to find out what the next action
should be as a result of the state change.  However, it wasn't resetting
the result cache between loops, so it was re-using the previous loop's
results each time, resulting in an infinite loop.  Reset the cache by
calling OvaleBestAction:StartNewAction().