Add simplistic translation of SimulationCraft's "incanters_flow_dir".
"incanters_flow_dir" is 1 if the flow is increasing and -1 if the flow is
decreasing, and probably 0 if the flow is unchanging at the extremes of
the cycle.  Translate this as "BuffDirection(incanters_flow_buff)" for
now, which is enough to get the current action prediction working.  It
doesn't currently compute the stacks properly over time as it relies on
updating the state from the current environment.