Make "check requirements" code more re-usable.
Create a registry of "requirement handlers" in OvaleData and provide two
new methods to make use of those handlers:

    OvaleData:CheckRequirements(tokenIterator, target)
    state:CheckRequirements(tokenIterator, target)

The required parameter "tokenIterator" is an iterator that returns a new
token every time it is evaluated.

This is designed to take a tokenIterator instead of a string so that
different separators can be used and the same handlers will still work.

Add new methods to retrieve spell info property with runtime checks:

    OvaleData:GetSpellInfoProperty(spellId, property, target)
    state:GetSpellInfoProperty(spellId, property, target)