Overhaul the way profiling statistics are gathered in Ovale.
Move profiling code from OvaleDebug to OvaleProfiler module.

Registering a module for profiling adds two new methods to the addon:
StartProfiling() and StopProfiling().  These two methods can be used at
the start and end of each function body to gather profiling statistics for
that function call.

Enabling and disabling profiling for a module via the Profiling options
panel replaces the StartProfilng and StopProfiling methods for that module
with functioning or do-nothing functions.