Warlock: Suggest Life Tap much earlier since we don't have infinite mana.
Mana is treated specially by Ovale as an infinite resource so that it's
not necessary to add mana costs to every spell.  That works in most cases
but fails in the case of warlocks that are supposed to use Life Tap in
their rotation to actively regenerate mana.  Move Life Tap much higher in
the rotation so that it gets used properly.