Not all stance keywords are compile-time parameter keywords.
Compile-time parameter keywords are for conditions that are evaluated
by OvaleCompile:EvaluateScript() before the script is run.  The "stance"
and "to_stance" parameter keywords are evaluated at run-time.

Before this change, "stance=NAME" was being ignored when set in SpellInfo,
which caused some spells to incorrectly be suggested even when not in the
required stance.

This fixes ticket 484 by @Shinzou01.