Add OvaleTotem module to keep state for player's totems.
OvaleTotem tracks everything that is considered to be a totem in-game.
This includes shaman totems, mage's Rune of Power and Prismatic Crystal,
druid Wild Mushroom, and monk statues.  Casting a spell that summons a
totem in the simulator will now actually summon the totem in the
simulator.

Modify script conditions to use the new state methods provided by
OvaleTotem.  RuneOfPowerRemaining() is deprecated and the Totem*()
conditions now directly take the ID of the spell that summons the totem.

Decorate totem spell descriptions in the default scripts with information
used by OvaleTotem.