Preliminary work on creating separate rotations from one SimC profile.
Spells are tagged as "main", "shortcd", or "cd" based on the length of
their cooldown and whether they invoke the GCD.  The tags are ordered as:
cd > shortcd > main.

The generated Ovale script now splits each action list into separate
functions corresponding to each of the tags.  Empty functions are not
generated and references to empty functions are removed from the final
script.

Enhance the generated script to use the standard four-icon layout of:

    [shortcd] [main] [aoe] [cd]

The "aoe" icon is toggled with a class- and specialization-specific
toggle.  This more closely matches the default Ovale scripts.