Add non-default tags to spells to match pre-existing default scripts.
Some spells need to be placed in particular icons instead of based on the
default tag as determined from CD and GCD by OvaleData.  Set them
explicitly in a separate section of the spells database for each class so
that the generated scripts more closely match the pre-existing default
Ovale scripts.