Fix state.nextCast to actually be the earliest start for next GCD spell.
This change also makes off-GCD spells suggested only when the GCD is not
on cooldown, which matches the behavior of SimulationCraft.

To allow a spell to be suggested during the global cooldown, it should be
tagged in the SpellInfo() with "offgcd=1".  This is useful for some spells
that should be cast right away, e.g., defensive cooldowns.

Fixes ticket 507 by @ShmooDude.