Apply auras at the correct time for channeled spells.
Channeled spells are instant and apply their effects when the spell is
cast.

Also remove the distinction between "on hit" and "after hit" for applying
the effects of a spellcast that is in flight.  Estimate the travel time of
the spell to determine when to apply the auras on the target.