Set the time of the next cast depending on the previous spell.
If the previous spell triggered the GCD, then wait till the next GCD cast
time to cast the next spell; otherwise, compute the next spell to cast at
the current time so that off-GCD spells that have priority can be
suggested immediately.

This matches the behavior of SimulationCraft where a whole batch of
off-GCD spells are cast when the next GCD spell is cast.