Translate SimulationCraft's text templating mechanism.
Text templating is basically preprocessor macro expansion.  Assume that
there is an order to the template definitions so that templates defined
later can use ones defined earlier in the file (not sure if that's
actually supported or true).  Do a straight textual substitution in the
action list before parsing it.

Fixes ticket 505 by @ShmooDude.