Differentiate between positional and named parameters.
The Ovale grammar allows for key=value parameters and if the key is a
small number, e.g., 1=value, then the node's parameter table could have an
overwritten positional parameter.

Split the positional and named parameters into separate tables so that
this is avoided.  Teach all parts of the code that references
"node.params" to use the correct table, either "node.positionalParams" or
"node.namedParams".

This changes the signatures for many functions due to having to pass both
parameter tables in place of just a single parameter table.