Avoid creating iterators at run-time using string.gmatch().
string.gmatch() returns a stateful iterator that consumes a small amount
of memory that must be garbage-collected after the iterator is discarded.
Ovale was calling string.gmatch() in its evaluation loop, which was
causing large amounts of memory to be consumed and discarded, which
triggered garbage-collection at a higher-than-normal rate.

Re-write the code paths so that there is no longer a need to parse a
string at run-time.  Comma-separated values are now parsed and split into
a table when the script is parsed in OvaleAST, and instead of using a
iterator over the CSV, a index into the CSV table is now used.