Raise the threshold to display a new action.
Ovale compares the start times of actions and selects the earliest one to
display.  If the new action is the same as the currently displayed action
and the start time is within COOLDOWN_THRESHOLD seconds, then Ovale
doesn't reset the cooldown sweep animation on the icon.

Raise COOLDOWN_THRESHOLD from 0.01 to 0.1, which should smooth out the
jitter/flicker in the display due to resource regeneration.

This addresses ticket 174 by @TuhMuffinMan.