Ensure that actions after an action-list function aren't suggested early.
Create a "post-conditions" function that can be used as a conditional for
the block of statements after calling an action-list function.  The
post-conditions conditional function ensures that the statements after the
action-list function call are never suggested earlier than they should be.

This fixes a SimulationCraft-to-Ovale translation error for profiles that
use "call_action_list".

Fix for ticket 547 by @Reason2012.