SimulationCraft can use "time_to_die" instead of "target.time_to_die".
"time_to_die" is both a character property and a time event in
SimulationCraft.  Silently translate all "time_to_die" to the equivalent
of "target.time_to_die".

Fixes problems in some scripts based on damage profiles where instead of
checking the target's time-to-die, it was using the player's instead,
which was clearly wrong.

Problem noted by @Wiljo.