Extend some OvaleEquipment methods to take a list of item slots.
The following methods are modified to accept a list of item slots instead
of a single or no parameters:

-  GetEquippedItem(...)
-  GetEquippedItemLevel(...)

The return value is the list of the item IDs in those slots, in order.

-  HasEquippedItem(itemId, ...)

Returns true if the item is in any of the listed slots, or check all of
the slots if none are given.