Add trinket buffs to the appropriate buff lists after script evaluation.
After the script has been evaluated, check for the equipped trinkets and
search for the associated buffs to add to the appropriate buff lists.

For example:

    Define(lucky_double_sided_coin 118876)
        ItemInfo(lucky_double_sided_coin buff=lucky_flip_buff)
    Define(lucky_flip_buff 177597)
        SpellInfo(lucky_flip_buff duration=20 stat=agility)

This causes the "Lucky" Flip buff to be added to the following buff lists:

  - trinket_proc_agility_buff
  - trinket_proc_agi_buff
  - trinket_proc_any_buff
  - trinket_stat_agility_buff
  - trinket_stat_agi_buff
  - trinket_stat_any_buff

The "buff" and "stat" properties can be comma-separated values to list
more than one buff or one stat.

Remove trinket buff lists from OvaleData since they are dynamically
managed during script evaluation.