Translate SimulationCraft's "using_apl" operand.
"using_apl" is meant to be set via a SimC command-line option in order to
choose the action list to run.  Translate this concept into an Ovale
dropdown menu option.

There is no good way to determine the valid APL choices from the profile,
so the translator looks for the "using_apl.<name>" operand, extracts the
APL name from the operand, and uses that APL name as an Ovale list item.