Add concept of refunding power after a spell is cast.
Change the PowerCost() and ComboPointCost() methods to return both the
power cost and the refunded power of a spell.  When applying the spell in
the simulator, also apply the power refund.

A spell that refunds power after it is cast can set "refund_<powerType>".
If refund_<powerType> is set to "cost" then it refunds all of the cost of
the spell.