Skip translating SimulationCraft's warrior "heroic_charge" action.
"Heroic Charge" means to move out of melee range and using Charge back to
the target in order to gain rage.  Skip translating this action for now as
the concepts needed to be checked (MH swing timer, combined cooldowns of
Heroic Leap and Charge based on talents, etc.) are too numerous to easily
code.