Automatically choose the correct SimulationCraft script by default.
Add "GetDefaultScriptName()" method to OvaleScripts.  This method
returns the name of the SimulationCraft script for a class/specialization
combination that can be used as a default script.

If the requested script is "Ovale", then map it to the correct
class/specialization-specific SimulationCraft script automatically by
using OvaleScripts:GetDefaultScriptName().

With these changes, there is no longer any need to manually glue together
all of the class-specific scripts to make one script that works for all
specializations.  Remove the jumbo and legacy scripts and split out the
resto shaman and resto druid scripts as "nerien" scripts to denote their
origin.