Reduce the number of stats that are snapshot into auras and spellcasts.
Save full snapshots into spellcasts but only save mastery and the base
damage multiplier into auras and discard the rest as snapshotting is
(mostly) dead.

Instead of reference-counting a snapshot table, just directly copy the
snapshot stats into the relevant places that need to save stats.

Remove script conditions to query stats that are no longer able to be
snapshot into an aura.