Remove the "wait" statement from the Ovale language.
It was too difficult to mix the imperative style of processing that "wait"
used with the time-span comparison style of processing that Ovale uses
otherwise within a single script.  Remove it to simplify the language.
The "wait" statement can be replaced with appropriate conditionals (see
OvaleSimulationCraft module for how it translates "wait").