OvaleTimeSpan now maintains its own table pool for time spans.
OvaleTimeSpan implements its own simple table pool instead of using
OvalePool so as to minimize dependencies and to make OvaleTimeSpan a more
standalone module. This should allow for time spans to be used in more
situations without having to worry about table ownership and also allows
for maximal re-use of table memory.

Remove the table pool for time spans from OvaleBestAction and update
timespan.t to check the size of the table pool.