Add a "Use tagged enemies" option to Ovale.
This new options determines the default behavior of the Enemies()
condition in Ovale scripts.  By default, it is false, which means that
Enemies() returns "activeEnemies".  If it is set to true, then it returns
"taggedEnemies" from the state machine.

Change contributed by @LunaEclipse with modifications by @jlam and closes
the following pullup requests on GitHub:

- Closes #2.
- Closes #3.