Ensure OnCheckBoxValueChanged() is run every time a checkbox is changed.
OnCheckBoxValueChanged() saves the checkbox state to the profile and
raises an event to trigger evaluating the script.  Ensure that this
private method is run regardless of how the checkbox state is changed,
whether through the UI or by directly calling SetCheckBox() or
ToggleCheckBox().

Fixes bug noted by @LunaEclipse and closes the following pullup requests
on GitHub:

- Closes #6.