Revert previous commit since Focus Fire uses the Frenzy stacks on the pet.
It's possible to get into a state where the player shows more stacks of
Frenzy than the pet does by dismissing and re-summoning the same pet.  In
this case, the status of the Frenzy buff between the player and the pet
is in an inconsistent state, but Focus Fire actually consumes the stacks
of Frenzy on the pet, regardless of how many stacks are displayed on the
player.

Also fix the name of the symbol used for the Frenzy buff in Ovale scripts.
This change makes the BM hunter script correctly suggest Focus Fire when
it reaches five stacks on the pet.