Arms and Protection update
The rage costs for Arms should be percent reductions and not an absolute value.  Currently you can only have percent reductions from an aura.  These reductions come from a talent.

Dragon Roar, 9 times out of 10, you'd want in in the main box, but I was reluctant to put it there with no way to change it (outside of a custom script).  What would be best is, similar to how movement abilities have a checkbox, you could have a just have a Dragon Roar checkbox.  I'm gonna do this in my custom script, might want to do it in the simcraft translator so it's automatic as well.  Not really too familiar with that yet so haven't tried.