Wrecking Ball fix
The fix is programmed specifically for Wrecking Ball/Whirlwind.  Possible there's a better place to put this but I wasn't having much luck elsewhere.  Put it there because that's the function conditions.lua calls.

Basically, the code checks if the aura is Wrecking Ball, then checks if the last (gcd) spell cast was Whirlwind.  If it was, and the aura started before the cast of Whirlwind, then sets the ending time of the aura to the starting time of the spellcast.