Separate cooldown and waiting on resources
Simulationcraft will stop evaluating the action list at the point where
a pool_resource line resolves to true.  Since Ovale evaluates the entire
action list before determining an action, the equivilent in Ovale is to
ignore the resource requirement when determining the best action.

Separates out the normal cooldown (or GCD) information from the extra
time required to pool resources for a spell.

pool_resource=1 will cause BestAction to ignore any additional time
needed for resource requirements.  This mimics Simulationcraft's
pool_resource,for_next=1 functionality.

extra_amount=# will increase the cost of an ability by that amount.  It
can be used with pool_resource=1 to mimic the
pool_resource,for_next=1,extra_amount=# functionality or on an ability
that consumes extra resources such as Ferocious Bite.

Fixed the red icon color and changed it to simply reference the now
separate actionResourcExtend.  If the nored parameter does not equal 1
and actionResourceExtend is greater than 0, then change the icon color
to red to indicate we are waiting on a lack of resources.

The simcraft translator should be updated to utilize this new
functionality on any pool_resource,for_next=1 or
pool_resource,for_next=1,extra_amount=# where # is a defined number.
Calculated numbers will still have to use if/unless logic but the only
script I've seen those in is the Rogue Subtlety script.