[DH][Vengeance] Update rotation from icy-veins.com
Reference:
http://www.icy-veins.com/wow/vengeance-demon-hunter-pve-tank-rotation-cooldowns-abilities

A. Single Target Rotation

The single target rotation for Vengeance Demon Hunters is based on the
following priority, and it is very straightforward, but keep in mind
that this does not account for your active mitigation, which we discuss
below.

The basic principle is to use up the Pain you generate through various
abilities so that you do not cap and waste it, while keeping enough
Pain in store to be able to use your active mitigation as needed.

1- Use Soul Carver on cooldown to generate Soul Fragments.
2- Use Soul Cleave to use up Pain and avoid capping it.
3- Use Immolation Aura on cooldown.
4- Use Felblade on cooldown (if you have taken this talent).
5- Use Fel Eruption on cooldown (if you have taken this talent).
6- Use Sigil of Flame.
7- Fill empty global cooldowns with Shear.

In addition to these abilities, you should also make use of Infernal
Strike (which is off the global cooldown) to deal additional damage.
We recommend always keeping at least 1 charge in reserve, however,
unless the fight requires no movement whatsoever, in which case you can
use up both charges.

B. Multiple Target Rotation

The multiple target rotation (again, not accounting for active
mitigation) is very similar to the single target one.

1- Use Soul Carver on cooldown to generate Soul Fragments.
2- Use Soul Cleave Icon Soul Cleave to use up Pain and avoid capping it.
3- Use Immolation Aura Icon Immolation Aura on cooldown.
4- Use Felblade on cooldown (if you have taken this
talent).
5- Use Fiery Brand on cooldown, only if you have the
Burning Alive talent.
6- Use Sigil of Flame Icon Sigil of Flame on cooldown.
7- Use Fel Eruption on cooldown (if you have taken
this talent).
8- Fill empty global cooldowns with Shear.

As in the single target rotation, you should also use Infernal Strike
against multiple targets.