Cooldown Swipe
Its hard to calculate the correct starting time for energy based abilities because energy is reported to us as discrete but we treat it as continuous.

This change instead updates the cooldown swipe, leaving it in the same percentage completion as before, but accelerating or decelerating it for the new end time.  This allows the cooldown swipe not to start over constantly but to keep the cooldown threshold at reasonable levels (would require a cooldown threshold of 250 ms to completely stop the swipe reset for energy based characters).