Blackout Kick/Tiger Palm Define are not for Windwalker only!
Hi,
Blackout Kick Define are not specified for Windwalker spec. Thus making it unusable in Windwalker spec.
I recommand specifying the cost as chi only in Windwalker spec. It also consume all the charges of teaching_of_the_monastery_buff.
Same goes for Tiger Palm. Tigger Palm does not require chi anymore, and seams not to apply tiger_power_buff since it is not in the game anymore. It does howerver apply teaching_of_the_monastery_buff.

Also can you explain to me the need/or not to specify a CD in the Define?

Thank you very much.

Mind that, I'am not a developper, I am just trying to help.